feat: add component for Header
