+++ src/components/header-left/header-left.tsx
@@ -1,0 +1,19 @@
+import { Link } from 'react-router-dom';
+
+function HeaderLeft(): JSX.Element {
+  return (
+    <div className="header__left">
+      <Link className="header__logo-link" to="/">
+        <img
+          className="header__logo"
+          src="img/logo.svg"
+          alt="6 cities logo"
+          width="81"
+          height="41"
+        />
+      </Link>
+    </div>
+  );
+}
+
+export { HeaderLeft };
