+++ src/components/header/header.tsx
@@ -1,0 +1,70 @@
+import { Link } from 'react-router-dom';
+import { useAppDispatch, useAppSelector } from '../../hooks';
+import { getAuthorizatinStatus } from '../../store/user-process/selectors';
+import { logoutAction } from '../../store/api-actions';
+import { AuthorizationStatus } from '../../const/auth';
+import { HeaderLeft } from '../header-left/header-left';
+
+function Header(): JSX.Element {
+  const authorizationStatus = useAppSelector(getAuthorizatinStatus);
+  const dispatch = useAppDispatch();
+  const email = localStorage.getItem('userEmail');
+  return (
+    <header className="header">
+      <div className="container">
+        <div className="header__wrapper">
+          <HeaderLeft />
+          <nav className="header__nav">
+            <ul className="header__nav-list">
+              {authorizationStatus === AuthorizationStatus.Auth ? (
+                <>
+                  <li className="header__nav-item user">
+                    <Link
+                      className="header__nav-link header__nav-link--profile"
+                      to="/favorites"
+                    >
+                      <div className="header__avatar-wrapper user__avatar-wrapper"></div>
+                      <span className="header__user-name user__name">
+                        {email}
+                      </span>
+                      <span className="header__favorite-count">3</span>
+                    </Link>
+                  </li>
+                  <li className="header__nav-item">
+                    <a
+                      className="header__nav-link"
+                      href="#"
+                      onClick={() => {
+                        dispatch(logoutAction());
+                      }}
+                    >
+                      <span className="header__signout">Sign out</span>
+                    </a>
+                  </li>
+                </>
+              ) : (
+                <>
+                  <li className="header__nav-item user">
+                    <a
+                      className="header__nav-link header__nav-link--profile"
+                      href="#"
+                    >
+                      <div className="header__avatar-wrapper user__avatar-wrapper"></div>
+                    </a>
+                  </li>
+                  <li className="header__nav-item">
+                    <Link className="header__nav-link" to="/login">
+                      <span className="header__login">Login</span>
+                    </Link>
+                  </li>
+                </>
+              )}
+            </ul>
+          </nav>
+        </div>
+      </div>
+    </header>
+  );
+}
+
+export { Header };
