+++ src/pages/favorites/favorites.tsx
@@ -1,59 +1,13 @@
-import { Link } from 'react-router-dom';
 import { FavoritesCard } from '../../components/favorites-card/favorites-card';
-import { useAppDispatch, useAppSelector } from '../../hooks';
-import { logoutAction } from '../../store/api-actions';
+import { useAppSelector } from '../../hooks';
 import { getFavorites } from '../../store/offers-favorite-data/selectors';
+import { Header } from '../../components/header/header';
 
 function Favorites(): JSX.Element {
   const favorites = useAppSelector(getFavorites);
-  const email = localStorage.getItem('userEmail');
-  const dispatch = useAppDispatch();
   return (
     <div className="page">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Link className="header__logo-link" to="/">
-                <img
-                  className="header__logo"
-                  src="img/logo.svg"
-                  alt="6 cities logo"
-                  width="81"
-                  height="41"
-                />
-              </Link>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a
-                    className="header__nav-link header__nav-link--profile"
-                    href="#"
-                  >
-                    <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-                    <span className="header__user-name user__name">
-                      {email}
-                    </span>
-                    <span className="header__favorite-count">3</span>
-                  </a>
-                </li>
-                <li className="header__nav-item">
-                  <a
-                    className="header__nav-link"
-                    onClick={() => {
-                      dispatch(logoutAction());
-                    }}
-                  >
-                    <span className="header__signout">Sign out</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
-
+      <Header />
       <main className="page__main page__main--favorites">
         <div className="page__favorites-container container">
           <section className="favorites">
